hal: move setps_util beside the code that links it
setps_util.c is named in hal/Submakefile for the python module and in
hal/utils/Submakefile for halcmd, halsh and halrmt, so hal/ was reaching
down into hal/utils/ for both a source file and a header.  It is shared HAL
support rather than a halcmd utility, so it moves up to hal/ and its three
users in hal/utils/ name the new path.

The same object is built into the same four targets.  With the edge gone
hal and hal/utils leave the include cycle, which is down to 8 directories.
