+++ src/hal/setps_util.c
