+++ src/hal/setps_util.h
