+++ src/hal/utils/Submakefile
@@ -2,13 +2,13 @@ HALCMDSRCS := \
     hal/utils/halcmd.c \
     hal/utils/halcmd_commands.cc \
     hal/utils/halcmd_main.c \
-    hal/utils/setps_util.c
+    hal/setps_util.c
 
 HALSHSRCS := \
     hal/utils/halcmd.c \
     hal/utils/halcmd_commands.cc \
     hal/utils/halsh.c \
-    hal/utils/setps_util.c
+    hal/setps_util.c
 
 ifneq ($(READLINE_LIBS),)
 HALCMDSRCS += hal/utils/halcmd_completion.c
@@ -33,7 +33,7 @@ endif
 	$(Q)$(CXX) $(LDFLAGS) -o $@ $^ $(READLINE_LIBS) -lfmt
 TARGETS += ../bin/halcmd
 
-HALRMTSRCS := hal/utils/halrmt.cc hal/utils/setps_util.c
+HALRMTSRCS := hal/utils/halrmt.cc hal/setps_util.c
 USERSRCS += $(HALRMTSRCS)
 
 ../bin/halrmt: $(call TOOBJS, $(HALRMTSRCS)) ../lib/liblinuxcnchal.so.0 ../lib/liblinuxcncini.so.1
